requirements.txt
playwright
click
pydantic
python-dotenv
nodriver
python-telegram-bot
apscheduler
